+++ packages/snap/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Enable P2TR (taproot) account creation alongside P2WPKH
+- Accept BIP-86 (purpose 86) derivation paths for taproot accounts
+- Include P2TR in account discovery
+
 ## [1.10.1]
 
 ### Fixed